16 lesson is done , todo - fix cart price

+++ src/ComponentsJSX/CartItem.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useDispatch } from "react-redux";
-import { addItem, removeItem } from "../redux/slices/cartSlice";
+import { addItem, removeItem, minusItem, clearItems } from "../redux/slices/cartSlice";
 
 const CartItem = ({ id, title, type, price, count, imageUrl }) => {
   const dispatch = useDispatch();
@@ -14,8 +14,15 @@ const CartItem = ({ id, title, type, price, count, imageUrl }) => {
   };
 
   const onClickMinus = () => {
-    dispatch(removeItem(id));
+    dispatch(minusItem(id));
   };
+
+  const onClickRemove = () => {
+    if (window.confirm('Вы уверены, что хотите удалить пиццу')) {
+      dispatch(removeItem(id));
+    }
+  };
+
 
   return (
     <div class="cart__item">
@@ -75,7 +82,10 @@ const CartItem = ({ id, title, type, price, count, imageUrl }) => {
         <b>{price * count} ₽</b>
       </div>
       <div class="cart__item-remove">
-        <div class="button button--outline button--circle">
+        <div
+          onClick={onClickRemove}
+          class="button button--outline button--circle"
+        >
           <svg
             width="10"
             height="10"
